backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
httpx>=0.27.0
pydantic-settings>=2.3.1
python-dotenv>=1.0.1
